# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waypoint Delivery Planning System API
  description: OpenAPI 3.1 specification for the Waypoint Group delivery planning system.
  contact:
    email: tech-triathlon@rootcode.io
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production API Base Endpoint
...
paths:
...
  /dashboard:
    get:
      operationId: getDashboard
      summary: Get dispatcher overview metrics and progress
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          description: Depot ID filter. Defaults if user has 1 depot. Returns 422 if missing when user has multiple depots, 403 if outside access.
          example: Peliyagoda
      responses:
        '200':
          description: Dispatcher dashboard metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
              example:
                date: "2025-08-01"
                depot_id: Peliyagoda
                business_now: "2025-07-31T14:00:00+05:30"
                cutoff:
                  cutoff_at: "2025-07-31T16:00:00+05:30"
                  minutes_remaining: 120
                  passed: false
                orders:
                  total: 120
                  by_brand:
                    Fresh: 80
                    Style: 25
                    Tech: 15
                  unassigned: 15
                  planned: 95
                  deferred: 10
                planning_progress:
                  planned: 95
                  total: 120
                vehicles:
                  total: 38
                  available: 34
                  in_workshop: 4
                  allocated: 28
                  loading: 5
                  active: 12
                alerts: []
                active_trips: []
        '403':
          description: Forbidden scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Missing required depot_id parameter for user with multiple depots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dispatch/queue:
    get:
      operationId: getDispatchQueue
      summary: Get unallocated order dispatch queue
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Unallocated dispatch queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
              example:
                - id: ORD-20250801-001
                  outlet_id: OUT004
                  delivery_date: "2025-08-01"
                  placed_at: "2025-07-31T14:00:00+05:30"
                  status: SUBMITTED
                  temp_requirement: chilled
                  order_units: 150
                  order_weight_kg: 180.0
                  order_volume_m3: 0.75
                  rolled_over: false
                  deferral_count: 0
        '403':
          description: Forbidden scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/generate:
    post:
      operationId: generatePlan
      summary: Generate daily dispatch plan using allocation engine
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePlanRequest'
            example:
              depot_id: Peliyagoda
              delivery_date: "2025-08-01"
              regenerate: false
      responses:
        '200':
          description: Generated plan run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRun'
              example:
                id: RUN-20250801-PEL
                depot_id: Peliyagoda
                delivery_date: "2025-08-01"
                status: OPEN
                version: 1
                generated_by: dispatcher@waypoint.test
                generated_at: "2025-07-31T14:05:00+05:30"
        '400':
          description: Non-operating day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans:
    get:
      operationId: listPlans
      summary: List dispatch plan runs
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: List of plan runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanRun'
  /plans/{plan_run_id}/trips:
    get:
      operationId: getPlanRunTrips
      summary: Get all trips for a specific plan run
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: plan_run_id
          in: path
          required: true
          schema:
            type: string
          example: RUN-20250801-PEL
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: List of trip details for the plan run
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripDetail'
              example:
                - trip:
                    id: TRIP-20250801-01
                    plan_run_id: RUN-20250801-PEL
                    delivery_date: "2025-08-01"
                    trip_no: 1
                    brand: Fresh
                    district: Colombo
                    depot_id: Peliyagoda
                    vehicle_id: VEH001
                    vehicle_type: truck
                    vehicle_temp: chilled
                    stop_count: 3
                    depart_time: "03:30"
                    plan_version: 1
                    status: CONFIRMED
                    weight_used_kg: 2400.0
                    weight_cap_kg: 3610.0
                    volume_used_m3: 12.5
                    volume_cap_m3: 19.4
                    est_minutes: 112
                    est_fuel_l: 12.4
                    time_budget_min: 270
                  stops:
                    - id: STOP-001
                      trip_id: TRIP-20250801-01
                      order_id: ORD-001
                      seq: 1
                      outlet_id: OUT004
                      outlet_name: Fresh Store Colombo 04
                      district: Colombo
                      dock_type: rear_dock
                      parking_constraint: normal
                      order_units: 50
                      order_weight_kg: 800.0
                      order_volume_m3: 4.1
                      window_open_time: "04:00"
                      window_close_time: "07:30"
                      eta: "04:15"
                      service_start_est: "04:15"
                      service_min: 15
                      status: PENDING
                      receipt_status: NONE
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Plan run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/validate:
    post:
      operationId: validatePlan
      summary: Validate plan against hard rules H1-H12
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePlanRequest'
            example:
              plan_run_id: RUN-20250801-PEL
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidatePlanResponse'
              example:
                valid: true
                violations: []
        '422':
          description: Constraint violations detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: CONSTRAINT_VIOLATION
                  message: Plan violates hard constraint H6 (volume capacity)
                  details:
                    violations:
                      - rule: H6
                        message: Volume capacity exceeded for vehicle VEH001
                        actual: 21.5
                        limit: 19.4
  /trips/{id}/orders:
    post:
      operationId: addOrderToTrip
      summary: Add order to a draft trip
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOrderToTripRequest'
            example:
              order_id: ORD-20250801-001
      responses:
        '200':
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '422':
          description: Constraint violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{id}/orders/{order_id}:
    delete:
      operationId: removeOrderFromTrip
      summary: Remove order from a draft trip
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
        - name: order_id
          in: path
          required: true
          schema:
            type: string
          example: ORD-20250801-001
      responses:
        '200':
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
  /orders/{id}/defer:
    post:
      operationId: deferOrder
      summary: Manually defer order with mandatory text reason
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: ORD-20250801-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeferOrderRequest'
            example:
              reason_text: Manual deferral due to dock maintenance at destination.
              note: Deferred by dispatcher.
      responses:
        '200':
          description: Deferral record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deferral'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: INVALID_TRANSITION
                  message: Cannot defer order in DELIVERED status
  /orders/{id}/requeue:
    post:
      operationId: requeueOrder
      summary: Return deferred order back to dispatch queue
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: ORD-20250801-001
      responses:
        '200':
          description: Requeued order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /trips/{id}/confirm:
    post:
      operationId: confirmTrip
      summary: Confirm draft trip plan for dispatch
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      responses:
        '200':
          description: Confirmed trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDetail'
              example:
                trip:
                  id: TRIP-20250801-01
                  plan_run_id: RUN-20250801-PEL
                  delivery_date: "2025-08-01"
                  trip_no: 1
                  brand: Fresh
                  district: Colombo
                  depot_id: Peliyagoda
                  vehicle_id: VEH001
                  vehicle_type: truck
                  vehicle_temp: chilled
                  stop_count: 1
                  depart_time: "03:30"
                  plan_version: 1
                  status: CONFIRMED
                  weight_used_kg: 180.0
                  weight_cap_kg: 3610.0
                  volume_used_m3: 0.75
                  volume_cap_m3: 19.4
                stops:
                  - id: STOP-20250801-001
                    trip_id: TRIP-20250801-01
                    order_id: ORD-20250801-001
                    seq: 1
                    outlet_id: OUT004
                    outlet_name: Fresh Colombo 03
                    district: Colombo
                    dock_type: street
                    parking_constraint: van_only
                    order_units: 150
                    order_weight_kg: 180.0
                    order_volume_m3: 0.75
                    window_open_time: "05:30"
                    window_close_time: "08:00"
                    eta: "06:15"
                    service_min: 16
                    status: PENDING
                    receipt_status: NONE
        '409':
          description: Invalid transition or plan changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Hard constraint violation blocking trip confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{id}/cancel:
    post:
      operationId: cancelTrip
      summary: Cancel a trip before departure
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      responses:
        '200':
          description: Cancelled trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '409':
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/close-run:
    post:
      operationId: closePlanRun
      summary: Close dispatch plan run and rollover unserved orders
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Closed plan run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRun'
        '409':
          description: Blocked because unassigned orders remain without formal deferrals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/summary:
    get:
      operationId: getPlansSummary
      summary: Get aggregate dispatch metrics and summary
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Summary of plan runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanRunSummaryResponse'
              example:
                plan_run_id: RUN-20250801-PEL
                status: OPEN
                version: 1
                trips: 12
                orders_served: 95
                orders_deferred: 10
                deferred_counts:
                  by_class:
                    UNAVOIDABLE: 4
                    CHOICE: 6
                  by_code:
                    CAPACITY_FULL: 6
                    NO_REEFER: 4
                capacity:
                  weight_used_kg: 32000.0
                  weight_cap_kg: 45000.0
                  volume_used_m3: 190.0
                  volume_cap_m3: 250.0
                reefer:
                  chilled_demand: 18
                  chilled_used: 14
                  chilled_available: 16
                fuel: []
                constraint_health:
                  compliant: true
                  violations: []
  /fuel:
    get:
      operationId: getFuelState
      summary: Get vehicle fuel quota usage and remaining litres
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Fuel ledger list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelLedger'
  /fleet:
    get:
      operationId: getFleetAvailability
      summary: Get fleet vehicle availability by date
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Fleet availability list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleAvailability'

  # ----------------------------------------------------
  # Dispatcher Monitoring & Exceptions Endpoints
  # ----------------------------------------------------
...
  /exceptions/{id}/decision:
    post:
      operationId: resolveExceptionDecision
      summary: Record dispatcher decision on driver exception (retry or skip)
      tags:
        - dispatcher
      x-role:
        - dispatcher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: EXC001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExceptionDecisionRequest'
            example:
              decision: skip
              note: Vehicle breakdown cannot be recovered in time for Fresh window.
      responses:
        '200':
          description: Updated exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        '422':
          description: Hard constraint violation (e.g. retrying a mall stop past window)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer authentication header. Role and scope are derived from the token.
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      description: Standard API error response envelope.

    # ----------------------------------------------------
    # Domain Model & DTO Schemas
    # ----------------------------------------------------
...
    TripDetail:
      type: object
      required:
        - trip
        - stops
      properties:
        trip:
          $ref: '#/components/schemas/TripCard'
        stops:
          type: array
          items:
            $ref: '#/components/schemas/StopDetail'
      description: Trip detail container with nested stop list (D23).
...
    GeneratePlanRequest:
      type: object
      required:
        - depot_id
        - delivery_date
      properties:
        depot_id:
          type: string
        delivery_date:
          type: string
          format: date
        regenerate:
          type: boolean
          description: If true, keeps existing locked (CONFIRMED) trips.
      description: Plan generation request payload.
...
    ValidatePlanRequest:
      type: object
      required:
        - plan_run_id
      properties:
        plan_run_id:
          type: string
      description: Plan validation request payload.
    ValidatePlanResponse:
      type: object
      required:
        - valid
        - violations
      properties:
        valid:
          type: boolean
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
      description: Plan validation result payload.
...